Began Refactoring Standalone Binary into Library
Moved all the core code powering the binary out of the main function
and into an analyzeDns function. Call `analyzeDns` with the parameters grabbed by
`flag.Parse()`. Instead of repetitive looping for min/avg/max/stddev/
calculations in the function body, I implemented Min(), Avg(), Max(),
and Std() methods on the DurationSlice type. This made some of the code
look prettier but it might turn out to be a bad idea.