Added Travis-CI Badge to the Readme
In addition to adding the Travis-CI Badge, I removed an extra `]`
trailing the end of the go report badge, and I changed the command in
the code sample to reflect that of the gif. I also grabbed the travis
linting rubygem and configured a local in .ruby-version to install it
and updated the .gitignore file to exclude it from the repo.